[Pal/lib] Fix snprintf() function signature

+++ Pal/lib/api.h
@@ -185,7 +185,7 @@ void fprintfmt (int (*_fputch)(void *, int, void *), void * f, void * putdat,
 void vfprintfmt (int (*_fputch)(void *, int, void *), void * f, void * putdat,
                  const char * fmt, va_list ap) __attribute__((format(printf, 4, 0)));
 
-int snprintf (char * buf, int n, const char * fmt, ...) __attribute__((format(printf, 3, 4)));
+int snprintf (char * buf, size_t n, const char * fmt, ...) __attribute__((format(printf, 3, 4)));
 
 /* Miscelleneous */
